backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic==2.12.4

langchain==0.3.7
langchain-core==0.3.31
langchain-community==0.3.7
langchain-text-splitters==0.3.2
langchain-google-genai==2.0.1

google-generativeai==0.8.5
chromadb==0.5.4
pypdf
unstructured[doc,docx]